+++ packages/tiny-toast/src/react/ui.tsx
@@ -2,6 +2,7 @@ import { Transition } from "@hiogawa/tiny-transition/dist/react";
 import { groupBy } from "@hiogawa/utils";
 import React from "react";
 import { TOAST_STEP } from "../core";
+import { cls, createPauseableTimeout } from "../utils";
 import type {
   ReactToastContainerOptions,
   ReactToastItem,
@@ -192,74 +193,4 @@ function ItemComponent({
       </div>
     </div>
   );
-}
-
-function cls(...args: unknown[]) {
-  return args.filter(Boolean).join(" ");
-}
-
-// pause-able setTimeout
-function createPauseableTimeout(
-  callback: () => void,
-  ms: number
-  // options?: { onDisposed?: () => void }
-) {
-  // TODO: handle ms === Infinity
-
-  type State =
-    | {
-        t: "stopped";
-      }
-    | {
-        t: "started";
-        startedAt: number;
-        stop: () => void;
-      }
-    | {
-        t: "disposed";
-      };
-
-  let state: State = { t: "stopped" };
-  let totalMs = 0;
-
-  function callbackWrapper() {
-    state = { t: "disposed" };
-    callback();
-    // options?.onDisposed?.();
-  }
-
-  return {
-    start: () => {
-      if (state.t === "stopped") {
-        state = {
-          t: "started",
-          startedAt: Date.now(),
-          stop: setupTimeout(callbackWrapper, ms - totalMs),
-        };
-      }
-    },
-    stop: () => {
-      if (state.t === "started") {
-        state.stop();
-        totalMs += Date.now() - state.startedAt;
-        state = { t: "stopped" };
-      }
-    },
-    dispose: () => {
-      if (state.t === "started") {
-        state.stop();
-      }
-      state = { t: "disposed" };
-      // options?.onDisposed?.();
-    },
-    state: () => state,
-  };
-}
-
-// destroy callback style api
-function setupTimeout(f: () => void, ms: number) {
-  let handle = setTimeout(f, ms);
-  return () => {
-    clearTimeout(handle);
-  };
 }
